# apps/idx-api/openapi.yaml
openapi: "3.0.3"
info:
  title: Idx Auth API
  description: |
    Authentication, authorization, user management, and chat thread API for Idx Chat.
    
    ## Response Envelope
    Every response follows a standardized format:
    ```json
    {
      "success": true | false,
      "requestId": "uuid-v4",
      "data": { ... },
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable message",
        "details": { ... }
      }
    }
    ```
    
    ## Authentication
    Most endpoints require a `Bearer` token in the `Authorization` header.
    Obtain tokens via `/auth/login`, `/auth/register`, or `/auth/google`.
  version: "0.1.0"
  contact:
    name: Idx Team
servers:
  - url: http://localhost:4000
    description: Development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      required: [success, requestId]
      properties:
        success: { type: boolean }
        requestId: { type: string, format: uuid }
        data: {}
        error:
          $ref: "#/components/schemas/ApiError"
    ApiError:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: {}
    SessionPayload:
      type: object
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer }
        user:
          $ref: "#/components/schemas/UserProfile"
        roles:
          type: array
          items: { $ref: "#/components/schemas/RoleRef" }
        permissions:
          type: array
          items: { type: string }
        permission_ids:
          type: array
          items: { type: integer }
    UserProfile:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        displayName: { type: string, nullable: true }
        avatarUrl: { type: string, nullable: true }
...
    RoleRef:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
...
paths:
...
  /auth/google:
    get:
      tags: [Auth]
      summary: Start Google OAuth flow
      parameters:
        - name: returnTo
          in: query
          schema: { type: string, default: "/" }
        - name: frontend
          in: query
          schema: { type: string }
      responses:
        "302": { description: Redirect to Google consent screen }
  /auth/google/callback:
    get:
      tags: [Auth]
      summary: Google OAuth callback (handled by Google)
      description: Redirects to frontend with exchange code
      parameters:
        - name: code
          in: query
          required: true
          schema: { type: string }
        - name: state
          in: query
          required: true
          schema: { type: string }
      responses:
        "302": { description: Redirect to frontend callback }
  /auth/exchange:
    post:
      tags: [Auth]
      summary: Exchange OAuth code for session tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
      responses:
        "200":
          description: Tokens issued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SessionPayload"
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                displayName: { type: string }
      responses:
        "201":
          description: User created, session returned
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SessionPayload"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email & password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SessionPayload"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SessionPayload"
        "401":
          $ref: "#/components/responses/InvalidToken"
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile & permissions
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current user data
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user: { $ref: "#/components/schemas/UserProfile" }
                          roles: { type: array, items: { $ref: "#/components/schemas/RoleRef" } }
                          permissions: { type: array, items: { type: string } }
                          permission_ids: { type: array, items: { type: integer } }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke refresh token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: Token revoked (no-op if already revoked)
  /auth/set-password:
    post:
      tags: [Auth]
      summary: Set password for OAuth-only account
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, minLength: 8 }
      responses:
        "200":
          description: Password set
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset
      description: Always returns 200 to avoid email enumeration. In dev mode, includes the reset token in the response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "200":
          description: Reset initiated (token included in dev mode)
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, minLength: 8 }
      responses:
        "200":
          description: Password reset successful
        "400":
          $ref: "#/components/responses/BadRequest"
  /auth/profile:
    patch:
      tags: [Auth]
      summary: Update own display name
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName: { type: string }
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          user: { $ref: "#/components/schemas/UserProfile" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/change-password:
    post:
      tags: [Auth]
      summary: Change own password
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [oldPassword, newPassword]
              properties:
                oldPassword: { type: string }
                newPassword: { type: string, minLength: 8 }
      responses:
        "200":
          description: Password changed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /auth/account:
    delete:
      tags: [Auth]
      summary: Delete own account
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Account deleted
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Admin ───────────────────────────────────────────
...